backend/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-cors-headers==3.5.0
psycopg2==2.8.6
python-socketio==5.0.3
